Switch to gorilla/mux
This allows us to use patterns in the Handler paths. Some of the GCP
tooling, like `gcloud`, might still contact the real metadata service
(or might use the GCE_METADATA_ROOT environment variable to determine
the host) before requesting a token from this service. We do not want to
prevent that or have to implement various endpoints. Instead, it is safe
to ignore the name of the Service Account for which tokens are
requested.

Signed-off-by: Dimitrios Karagiannis <dhkarag@gmail.com>