Implement operator
Watches serviceaccounts and grants access in vault to the role arn
requested in the vault.uw.systems/aws-role annotation. Removes them when
the annotation is removed or the serviceaccount is deleted.

On startup the operator will go through all the items in vault and
remove any dangling references to serviceaccounts that no longer exist
in the cluster. This might happen if the operator is down for a period
of time.

The objects in vault are named `<prefix>_aws_<namespace>_<name>`. The
prefix allows us to insert the environment into the name of the object,
which we do in gcp. Putting 'aws' in there will make the logic of adding
gcp or other cloud providers much easier.

I've modified the flags and moved to a subcommand + flags model. I found
the list of flags hard to reason about from the usage and I think it's a
lot clearer now. It does involve some duplication of flags between
subcommands but I don't really mind that.

What this doesn't include is validation that the serviceaccount is
permitted to assume the role it has requested. That will be coming in a
future commit.