Validate that the operator prefix does not contain an underscore (#12)
We use _ to separate elements in the names of the items the operator
writes to vault. The operator needs to be able to split these items back
into their constituent parts reliably in order to garbage collect items
in vault and a prefix with an underscore would make this impossible.