add structured logging with zap
By default the controller-runtime manager logs with a null logger, so in
order to capture log messages internal to the manager you need to set a
go-logr compatible logger.

Zap is the common one suggested by controller-runtime, so I've gone with
that.

I've amended the sidecars to use it as well as the operator for the sake
of consistency.